Refactor the main file for readability
Create a module with the algorithm to solve the expressions, in order
to reduce the code lines.

As every time the program resolve the expression, the algorithm has to
be used multiple times, it better be in a function to make it reusable.

Create the calculator.rs module

Create the compute() function which uses the methods to calculate the
expression