Add some validations
Add two functions to validate the stack and valid the case any number
divided by zero.

As the program uses a static stack, then exist the case when the stack
is full or empty and it isn't possible excute the pop and push.

In the module "stack", there are two methods which return a value in the
respective case, add two functions to excute a panic, which shows an
error, to stop the program.

In order to validate when a number is divided by zero, add an if
satement to know if the denominator is zero.

Add in module calculator.rs
push()
pop()