Create the Tree Data Structure
Create a binary three and add the tree methods.

That structure is used to transform the expression order. The
notation, which we used, is called infix and the notation, which the
stack needs to compute the expression, is called postfix.

Create the tree module and implement the methods. When the tree has the
data, you should use the method post_order() and use the method
compute() which computes a posfix expression.